Rename snippet related components and more descriptive comments
